rename: 공통 버튼 폴더 생성
